Added 'delete bucket' and moved sunburst logic to its own vue component
