+++ src/visualizations/coloring.js
@@ -4,7 +4,7 @@ export default {
     },
     afkstatus: {
         key: "status",
-        colors: {"afk": "#CCC", "not-afk": "#0F4", "hibernating": "FF6"}
+        colors: {"afk": "#CCC", "not-afk": "#0F4", "hibernating": "#FF6"}
     },
     "currently-playing": {
         key: "title"
